:memo: Add docs for contributing and private items (#26)
Hello! I took some time between my last PR and this one because I was documenting everything on the code 😍 .

This is the list of contributions, I would like to hear the thoughts of @fdncred and @amtoine about them before merging this PR:

- `docs/CONTRIBUTION.md` added. I explained the why and the how to contribute, and appended a list of general guidelines about the philosophy of the project:
> - Everything should be explained: rust docs, comments, drawings, pick what makes you comfortable, but it is important to make it clear. There will always be some new guy or gal into the project we want to welcome 😄.
> - Use clear variable names and try to avoid confusing abbreviations. Think that your peers may not be fully fluent in english 💬.

- I added a few clippy lints from `clippy:pedantic`, [here](https://github.com/nushell/nufmt/pull/26/commits/4b97279ad6822b916dba66a3ea1272387d0cc371#diff-b1a35a68f14e696205874893c07fd24fdb88882b47c23cc0e0c80a30c7d53759R6) are them: Clippy will give a warning if one if them is not compliant. 

I added these rules because I would like to improve the quality of upcoming contributions. I'm not sure if this is asking too much from the people who want to be involved in the project, as some of the rules may be a bit annoying. Do you see it doable? Do you think we should delete some or all of the clippy lints?