rename the library into `nu_formatter`
the goal of this commit is to make a clear distinction between
the library which does all the heavy lifting and the binary
application which mainly just parses the arguments and calls the
library.

because i think the application is clearly called `nufmt` i
propose here to change the name of the library to `nu_formatter`.
i also thought about `nufmt_lib`!

this has the huge benefit of making two separate items in the
`cargo doc` instead of two with the same `nufmt` name :/

```
cargo doc --document-private-items --bin nufmt --open
```
directly opens the doc on the application page with the help and
the usage, which i find bette :)