Update nufmt (#75)
## Description of changes

This is my attempt to get `nufmt` working better. I've been working on
it for a while. The readme.md describes more but it's some rust changes
and a lot of testing script files with inputs and expected outputs.

Give it a try and let me know if it works for you too.